Add files via upload
Date created
2021-01-18

Project title
Exploring US Bikeshare Data

Description
In this In this project, Python code is written to import U.S. bike share data and answer interesting questions about it by computing descriptive statistics. Also, a script is written that takes in raw input to create an interactive experience in the terminal to present these statistics.

The project displayed bike sharing data for three major cities —Chicago, New York City, and Washington. An interactive environment was created where a user chooses the data and filter for a dataset to analyze.

Files used
- bikeshare.py
- chicago.csv
- new_york_city.csv
- washington.csv

How to use the bikeshare interactive environment
- Run bikeshare.py file using Git Bash (for Windows)
- Answer a few questions to select your filter (i.e., city, month, day)
- The program will calculate the statistics and display system usage for your chosen filters, such as popular times of travel, popular station and trip, trip duration, user info, etc.