Add files via upload
In this project, SQL is used to explore a database related to movie rentals. SQL code is written to run SQL queries and answer interesting questions about the database. Finally,  visualizations are prepared to showcase the output of the queries.
The database is queried to gain an understanding of the customer base, such as what the patterns in movie watching are across different customer groups, how they compare on payment earnings, and how the stores compare in their performance.